requirements.txt
PyYAML==6.0.2
pip==24.2
pygame==2.6.0
screeninfo==0.8.1
setuptools==74.1.2
wheel==0.44.0
pygame-widgets==1.1.5